+++ SellingRentingCarsSystem.API/Models/PowerTrain.cs
@@ -5,11 +5,11 @@ public class PowerTrain
     public string Id { get; set; } = Guid.CreateVersion7().ToString();
     public string PowerTrainType { get; set; } = string.Empty;
     public int HorsePower { get; set; }
-    public decimal Torque { get; set; }
-    public decimal? CombinedRangeMiles { get; set; }
-    public decimal? ElectricOnlyRangeMiles { get; set; }
+    public int Torque { get; set; }
+    public float? CombinedRangeMiles { get; set; }
+    public float? ElectricOnlyRangeMiles { get; set; }
     public string? ChargePortID { get; set; }
-    public decimal? BatteryCapacityKWh { get; set; }
+    public float? BatteryCapacityKWh { get; set; }
     public string? FuelDeliveryID { get; set; }
     public string? FuelTypeID { get; set; }
     public string? AspirationID { get; set; }
